dbeaver/pro#2555 Redesign to handler, moved to standalone
