--- plugins/org.jkiss.dbeaver.registry/OSGI-INF/org.jkiss.dbeaver.databaseinfo.service.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.jkiss.dbeaver.service.DataBaseInfoService">
-   <implementation class="org.jkiss.dbeaver.registry.service.DataBaseInfoServiceImpl"/>
-   <service>
-      <provide interface="org.jkiss.dbeaver.registry.DataBaseInfo"/>
-   </service>
-</scr:component>
