+++ plugins/org.jkiss.dbeaver.registry/build.properties
@@ -3,7 +3,6 @@ bin.includes = .,\
                META-INF/,\
                OSGI-INF/,\
                plugin.xml,\
-               schema/,\
-               OSGI-INF/org.jkiss.dbeaver.databaseinfo.service.xml
+               schema/
 source.. = src/
 output.. = target/classes/
