+++ plugins/org.jkiss.dbeaver.ui.app.standalone/src/org/jkiss/dbeaver/ui/app/standalone/ApplicationWorkbenchWindowAdvisor.java
@@ -60,9 +60,6 @@ import java.util.StringJoiner;
 
 public class ApplicationWorkbenchWindowAdvisor extends IDEWorkbenchWindowAdvisor implements DBPProjectListener, IResourceChangeListener {
 
-    private static final String OUTPUT_DATABASES_JSON = "supported.databases.json"; //$NON-NLS-1$
-    private static final String CLI_DATABASE_LOG = "-databaseLog"; //$NON-NLS-1$
-
     private static final Log log = Log.getLog(ApplicationWorkbenchWindowAdvisor.class);
     
     // Eclipse fonts
@@ -381,23 +378,6 @@ public class ApplicationWorkbenchWindowAdvisor extends IDEWorkbenchWindowAdvisor
         if (isRunWorkbenchInitializers()) {
             // Open New Connection wizard
                 initWorkbenchWindows();
-        }
-        logSupportedDataBases();
-    }
-
-    /**
-     * Log supported database details for arg -databaseLog
-     */
-    private void logSupportedDataBases() {
-        String[] applicationArgs = Platform.getApplicationArgs();
-        if (applicationArgs != null && Arrays.asList(applicationArgs).contains(CLI_DATABASE_LOG)) {
-            DataBaseInfo databaseInfoService = RuntimeUtils.getBundleService(DataBaseInfo.class, true);
-            if (databaseInfoService != null) {
-                DBeaverApplication instance = DBeaverApplication.getInstance();
-
-                databaseInfoService.publishDataBaseInfo(
-                    instance.getDefaultWorkingFolder().resolve(OUTPUT_DATABASES_JSON));
-            }
         }
     }
